Аудит v2 волна 3: все MEDIUM и LOW закрыты
M-01: удалены 9 stub-файлов (crypto, dns, routing, watchdog).
M-02: blake2s_hmac = keyed BLAKE2s (Noise spec).
M-03: nonce uint64_t LE для WireGuard AEAD.
M-04: time(NULL) кэш once per tick.
M-05: stderr пропуск в daemon mode.
M-06: DNS cache probe max 16.
M-07: dns_rules ранний выход подтверждён.
M-08: stdbool.h в dns_cache.h.
M-09: ss_recv tmp буфер при truncation.
M-10: ntp send() цикл.
M-11: xhttp getrandom + O_CLOEXEC.
M-12: health[64] вместо 8.
L-03: DISPATCHER_MAX_HEALTH=64.
L-04: tproxy UDP buf на heap.
L-05: AWG init_pkt[1536].
L-06: AWG pkt[1536] + len cap 1420.
Аудит v2: 34/43 закрыто.

+++ core/include/crypto/noise.h
@@ -20,8 +20,8 @@ typedef struct {
     /* Результат handshake */
     uint8_t  send_key[32];
     uint8_t  recv_key[32];
-    uint32_t send_counter;
-    uint32_t recv_counter;
+    uint64_t send_counter;
+    uint64_t recv_counter;
     uint32_t local_index;
     uint32_t remote_index;
     bool     handshake_complete;
